refactor(*): Flatten directory structure
