+++ app/js/skillToggle.js
