+++ app/styles/badges.scss
