+++ app/styles/leaderboard.scss
