+++ server/app.js
@@ -16,10 +16,10 @@ const Mongo = require('./db');
 const app = express();
 
 // view engine setup
-app.set('views', path.join(__dirname, '..','client', 'views'));
+app.set('views', path.join(__dirname, 'views'));
 app.set('view engine', 'pug');
 
-app.use(favicon(path.join(__dirname, '..', '..', 'public', 'favicon.ico')));
+app.use(favicon(path.join(__dirname, '..', 'public', 'favicon.ico')));
 app.use(morgan('combined'));
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({
