+++ server/badge.js
