+++ server/views/badges.pug
